requirements.txt
SQLAlchemy==2.0.31
psycopg2==2.9.9
python-dotenv==1.0.1